fix(storage): resolve task canonical hash mismatch during replay
Root cause: Hash computation was inconsistent between write and replay paths
due to different JSON structure for acceptance criteria and other nested objects.

Changes:
- Add JSON normalization in TaskToSnapshot() to ensure consistent structure
- Add normalizeStateMap() call before storing snapshots in saveVersionUnlocked()
- Exclude all timestamp fields (createdAt, updatedAt, completedAt, archivedAt)
  from canonical hash computation
- Triple JSON round-trip in hashSnapshot() to ensure deterministic encoding
- Fix test assertion to expect float64 instead of int after JSON round-trip

This ensures snapshots have identical structure whether they come from:
1. Write path: strongly-typed structs → TaskToSnapshot() → normalized map
2. Replay path: JSONL unmarshal → already normalized map

Fixes TestCLI_ReopenWorkflow hash mismatch at revision 8.

+++ internal/storage/history_jsonl_test.go
@@ -101,7 +101,8 @@ func TestTaskHistoryReplayPreservesZeroValueDeltas(t *testing.T) {
 	if archived, ok := snapshot["archived"].(bool); !ok || archived {
 		t.Fatalf("replayed archived = %#v, want false", snapshot["archived"])
 	}
-	if spent, ok := snapshot["timeSpent"].(int); !ok || spent != 0 {
+	// After JSON round-trip, numbers become float64
+	if spent, ok := snapshot["timeSpent"].(float64); !ok || spent != 0 {
 		t.Fatalf("replayed timeSpent = %#v, want 0", snapshot["timeSpent"])
 	}
 	if _, ok := snapshot["description"]; ok {
